% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quotes.R
\name{get_categories}
\alias{get_categories}
\title{Return the categories from the API}
\usage{
get_categories()
}
\value{
Vector of comment categories
}
\description{
Return the categories from the API
}
\examples{
get_catgories()
}
