% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quotes.R
\name{get_quote}
\alias{get_quote}
\title{Return a quote}
\usage{
get_quote(category = NULL)
}
\arguments{
\item{category}{One of the quote categories as a string}
}
\value{
Quote as text
}
\description{
Return a quote
}
\examples{
get_quote()
get_quote("funny")
}
